+++ test/cpp_extensions/open_registration_extension/torch_openreg/csrc/CMakeLists.txt
@@ -9,16 +9,10 @@ add_library(${LIBRARY_NAME} SHARED ${SOURCE_FILES})
 target_link_libraries(${LIBRARY_NAME} PRIVATE torch_cpu_library openreg)
 target_include_directories(${LIBRARY_NAME} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
-if (WIN32)
-    target_compile_definitions(${LIBRARY_NAME} PRIVATE OPENREG_EXPORTS)
-endif()
+set(CMAKE_CXX_VISIBILITY_PRESET hidden)
 
-if(WIN32)
-    install(TARGETS ${LIBRARY_NAME} 
-        ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
-        LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
-        RUNTIME DESTINATION ${CMAKE_INSTALL_LIBDIR}
-    )
-else()
-    install(TARGETS ${LIBRARY_NAME} LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR})
-endif()
+install(TARGETS ${LIBRARY_NAME} 
+    ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
+    LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
+    RUNTIME DESTINATION ${CMAKE_INSTALL_LIBDIR}
+)
