+++ test/cpp_extensions/open_registration_extension/torch_openreg/csrc/runtime/OpenRegFunctions.h
@@ -1,13 +1,9 @@
 #pragma once
 
 #ifdef _WIN32
-  #ifdef OPENREG_EXPORTS
-    #define OPENREG_API __declspec(dllexport)
-  #else
-    #define OPENREG_API __declspec(dllimport)
-  #endif
+  #define OPENREG_EXPORT __declspec(dllexport)
 #else
-  #define OPENREG_API
+  #define OPENREG_EXPORT __attribute__((visibility("default")))
 #endif
 
 #include <c10/core/Device.h>
@@ -17,10 +13,10 @@
 
 namespace c10::openreg {
 
-OPENREG_API c10::DeviceIndex device_count() noexcept;
-OPENREG_API c10::DeviceIndex current_device();
-OPENREG_API void set_device(c10::DeviceIndex device);
+OPENREG_EXPORT c10::DeviceIndex device_count() noexcept;
+OPENREG_EXPORT c10::DeviceIndex current_device();
+OPENREG_EXPORT void set_device(c10::DeviceIndex device);
 
-OPENREG_API DeviceIndex ExchangeDevice(DeviceIndex device);
+OPENREG_EXPORT DeviceIndex ExchangeDevice(DeviceIndex device);
 
 } // namespace c10::openreg
