+++ test/cpp_extensions/open_registration_extension/torch_openreg/setup.py
@@ -19,6 +19,19 @@ BASE_DIR = os.path.dirname(os.path.realpath(__file__))
 RUN_BUILD_DEPS = any(arg in {"clean", "dist_info"} for arg in sys.argv)
 
 
+def check_env_flag(name, default = ""):
+    return os.getenv(name, default).upper() in ["ON", "1", "YES", "TRUE", "Y"]
+
+
+if "CMAKE_BUILD_TYPE" not in os.environ:
+    if check_env_flag("DEBUG"):
+        os.environ["CMAKE_BUILD_TYPE"] = "Debug"
+    elif check_env_flag("REL_WITH_DEB_INFO"):
+        os.environ["CMAKE_BUILD_TYPE"] = "RelWithDebInfo"
+    else:
+        os.environ["CMAKE_BUILD_TYPE"] = "Release"
+
+
 def make_relative_rpath_args(path):
     if IS_DARWIN:
         return ["-Wl,-rpath,@loader_path/" + path]
@@ -54,7 +67,7 @@ def build_deps():
         ".",
         "--target",
         "install",
-        "--config", "Release",
+        "--config", os.environ["CMAKE_BUILD_TYPE"],
         "--",
     ]
 
@@ -83,11 +96,44 @@ class BuildClean(clean):
 def main():
     if not RUN_BUILD_DEPS:
         build_deps()
-    
-    if sys.platform == "win32":
-        extra_compile_args = ["/W3"]
+
+    if IS_WINDOWS:
+        # /NODEFAULTLIB makes sure we only link to DLL runtime
+        # and matches the flags set for protobuf and ONNX
+        extra_link_args: list[str] = ["/NODEFAULTLIB:LIBCMT.LIB"]
+        # /MD links against DLL runtime
+        # and matches the flags set for protobuf and ONNX
+        # /EHsc is about standard C++ exception handling
+        extra_compile_args: list[str] = ["/MD", "/FS", "/EHsc"]
     else:
-        extra_compile_args = ["-g", "-Wall", "-Werror"]
+        extra_link_args = []
+        extra_compile_args = [
+            "-Wall",
+            "-Wextra",
+            "-Wno-strict-overflow",
+            "-Wno-unused-parameter",
+            "-Wno-missing-field-initializers",
+            "-Wno-unknown-pragmas",
+            # Python 2.6 requires -fno-strict-aliasing, see
+            # http://legacy.python.org/dev/peps/pep-3123/
+            # We also depend on it in our code (even Python 3).
+            "-fno-strict-aliasing",
+        ]
+
+    if os.environ["CMAKE_BUILD_TYPE"] == "Debug":
+        if IS_WINDOWS:
+            extra_compile_args += ["/Z7"]
+            extra_link_args += ["/DEBUG:FULL"]
+        else:
+            extra_compile_args += ["-O0", "-g"]
+            extra_link_args += ["-O0", "-g"]
+    elif os.environ["CMAKE_BUILD_TYPE"] == "RelWithDebInfo":
+        if IS_WINDOWS:
+            extra_compile_args += ["/Z7"]
+            extra_link_args += ["/DEBUG:FULL"]
+        else:
+            extra_compile_args += ["-g"]
+            extra_link_args += ["-g"]
 
     ext_modules = [
         Extension(
